address_cache

Handles caching of already-processed addresses so that they're only processed once

Usage

address_cache(
  address_stream,
  cache_name,
  .function,
  key_cols = address_cols,
  db_name = tessilake::cache_primary_path("address_stream.sqlite", "stream"),
  ...
)

address_cache_chunked(
  address_stream,
  cache_name,
  .function,
  key_cols = address_cols,
  db_name = tessilake::cache_primary_path("address_stream.sqlite", "stream"),
  parallel = rlang::is_installed("furrr"),
  n = 100,
  ...
)

Arguments

address_stream

data.table of addresses

cache_name

name of cache file

.function

function to be called for processing, is sent address_stream[address_cols] and additional parameters.

key_cols

address_cols

db_name

path to sqlite database, defaults to tessilake::cache_primary_path("address_stream.sqlite","stream")

...

additional options passed to .function

parallel

boolean whether to run chunks in parallel, defaults to TRUE when furrr is installed.

n

integer chunk size

Value

data.table of addresses processed

Functions

address_cache_chunked(): Parallel wrapper around address_cache using furrr::furrr and progressr::progressr