address_stream

Create a data.table stream of addresses and timestamps, drawing from Tessitura TA_AUDIT_TRAIL and T_ADDRESS data, geocoding using the tidygeocoder package. Tries each address up to six times, using libpostal parsing, street1, and street2, and the US census and openstreetmap geocoders. Appends census demographic and aggregate income data and iWave income data.

Usage

address_stream(...)

address_stream_build(...)

Arguments

...

parameters passed on to address_create_stream

Value

data.table of addresses and additional address-based features

Functions

address_stream_build(): build the address_stream by combining data from address_create_stream, address_parse, address_geocode, address_census, and iWave data from tessilake::read_tessi

Note

Uses future::future for parallel processing and progressr::progressr for progress tracking