survey_cross

survey_cross

Usage

survey_cross(survey_stream, ...)

Arguments

survey_stream

data.table of survey data

...

one or more regular expressions used to locate the questions to combine

Value

data.table of survey data, merged on customer_hash so that each question/answer is a separate column, labelled question.1, answer.1, question.2, answer.2, etc. based on the position of the selecting regular expression in ...